[TASK] Simplify package, removing optional models and plugins
This change mainly removes features to simplify this package.
In detail it:

* ..removes the "Frontend user profile" plugin
* ..removes the custom ``User`` model, controller and the ``FrontendUserService``
* ..removes the custom ``frontenduser`` CommandController
* ..renames the ``LoginController`` to ``AuthenticationController``

Background:

This package used to come with a custom FrontendUser domain model and
a "profile editing" plugin for demonstration purposes, but this had two
major drawbacks:

* It prevented the package from being used for other models without
  adjustments
* It mislead people into using the custom model as base and extend
  from there, which contradicts a core DDD concept that suggests to
  build your models as close as possible to the actual domain

Therefore we decided to simplify the package and remove custom
models/plugins. Of course you're free to use it with your custom
User models or even fork the original ``User`` model of the 1.0 branch.

Resolves: `NEOS-1340 <https://jira.typo3.org/browse/NEOS-1340>`_